+++ native/BANG_FX/src/BANG_StrokePiPL.r
@@ -22,7 +22,7 @@ resource 'PiPL' (16000) {
 #endif
 		AE_PiPL_Version { 2, 0 },
 		AE_Effect_Spec_Version { PF_PLUG_IN_VERSION, PF_PLUG_IN_SUBVERS },
-		AE_Effect_Version { 622593 /* 1.3.0 build 1 */ },
+		AE_Effect_Version { 655361 /* 1.4.0 build 1 */ },
 		AE_Effect_Info_Flags { 0 },
 		/* PF_OutFlag_DEEP_COLOR_AWARE | PF_OutFlag_I_EXPAND_BUFFER */
 		AE_Effect_Global_OutFlags { 0x02000200 },